Struct Import Export (#39)
* Writing char array

* Basic Read Write

* Adding Writing For Generic Types

* Update serializer.h

* Struct Read Write Func

Added Writing for char, short, int, long, double, string.
Added Reading for char,short, int, long, double, string.

* Added Generic List Support

Added Support for lists with generic types.
Norm Support.

* Update serializer.h

* Relying on a header pos

Added header pos for better future compatiblity.

* Endian Support

Added Support for both endians

* Update read_list_gen.c

Now support endians

* Added Error Messages

Added allocation error to read functions.

* More Binary Struct Wrappers

* Made Reading Beautiful

* Added Define Path

* Update main.c

* Update detect_endian.c

* Update read_list_gen.c

* Delete help.txt

* Changing Long to t_uint64

* Fixing Typedefs

* Fixing Suggestions

* Making Header Support Endians

+++ libft/src/list/ft_putnumlst.c
@@ -17,11 +17,11 @@
 
 t_bool	ft_putnumlst(const t_list *lst)
 {
-	intmax_t	*n;
+	int	*n;
 
 	if (lst == NULL)
 		return (ft_putstr(NULL));
-	n = (intmax_t *)lst->content;
+	n = (int *)lst->content;
 	if (ft_putnum(*n) == FALSE)
 		return (FALSE);
 	if (ft_putchar('\n') == FALSE)
